Move example code to README (#5)
The code exampleServer.go was added so that dep can correctly identify this repo as a Go repo. However, the file is not really used and the example code is better put into the README as a reference.

This is replaced with a file named dep.go and a comment explaining the purpose of the file.